fix: styles on main page
