+++ src/App.css
@@ -41,7 +41,7 @@
   color: #888;
 }
 
-.top-controls input {
+.top-controls #switch {
   height: 0;
   width: 0;
   visibility: hidden;
